requirements.txt
opencv-python
numpy
cvlib
Pillow
matplotlib
